TalentMatchAI/requirements.txt
# ==============================
# Core API
# ==============================
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9

# ==============================
# RAG / Vector DB
# ==============================
chromadb==0.5.0
numpy==1.26.4
tqdm==4.66.2

# ==============================
# OpenAI (Compatible with httpx 0.27)
# ==============================
openai==1.30.1
httpx==0.27.0

# ==============================
# Resume Parsing
# ==============================
pypdf==6.7.1
python-docx==1.1.0

# ==============================
# Gmail Integration
# ==============================
google-api-python-client==2.130.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ==============================
# Optional (Recommended for Production)
# ==============================
orjson==3.10.3
